+++ NEWS.md
@@ -9,6 +9,9 @@ Major rewrite of the download and synchronization code, with some **breaking cha
 * Column names in the database are always, and with raw files by default, normalized by lower-casing and replacing all spaces with underscores, e.g. "Event ID" to "event_id". This is to make working in R and SQL with the events data easier, i.e. no need to backquote names. 
 
 * Change `read_icews` to also work if a database backend is used, previously only for reading from files. 
+* Change `cameo_codes` data to include columns for quad and penta categories. 
+* Add `icews_to_gwcode` to translate ICEWS country names to their respective Gleditsch & Ward country code. 
+
 
 # icews 0.1.0